feat(shell): C3 — device facts (dppx + prefers-color-scheme) shell→content delivery (#415)
Delivers the two per-window device facts — `window.devicePixelRatio` / `@media (resolution)` (dppx) and `prefers-color-scheme` — from the shell to every content thread, with a live canonical media consumer (boa routed through `elidex_css::media::evaluate`). Closes the F3 `ScaleFactorChanged` X11 slot and completes the PR-C axis-3 producer-delivery breadth (C1 size / C2 geometric / C3 device-fact).

Producer: `ViewportCell` generalized to `{ size, seq, facts, facts_seq }`; one redraw-top chokepoint (`refresh_viewport_cell`) recomputes placement + publishes atomically, called before the content-message drain AND before any tab spawn so a construction-input spawn never reads a stale cell. `SetViewport` carries the settled facts so a frame changing both size and facts delivers one atomic content re-eval (no inconsistent intermediate MQL `change`); `SetDeviceFacts` handles facts-only frames; each carries its own staleness generation. dppx is carried as lossless `f64` (CSSOM View `devicePixelRatio` is a `double`); the media comparison aligns to the cssparser-`f32` query precision in the engine-independent evaluator (`elidex_css::media::eval`). egui chrome gets DPI/theme events via a `forward_to_egui` helper.

Codex `/external-converge`: 5 rounds → TERMINAL (R4+R5 dry). 8 in-scope fixes (facts_seq build-vs-broadcast staleness, egui DPI/theme chrome, iframe build-time device-fact seed, atomic size+facts delivery, publish-before-drain spawn freshness, dppx f64 precision, spawn-cell-freshness root fix, eval.rs precision-alignment layering). 2 self-root-checks + 2 cumulative design re-gates (the R3 re-gate caught a layering slip — precision-alignment misplaced in the boa producer — corrected to the engine-independent evaluator). Carves (genuine cross-PR scope boundaries): `#11-media-css-values-fidelity` (CSS-cascade `<style>@media` rendering — engine-independent elidex-style layer), `#11-iframe-device-fact-live-propagation` (live propagation to already-loaded iframes — edge-dense new IPC mechanism).

+++ crates/script/elidex-js-boa/src/globals/window/screen.rs
@@ -134,8 +134,9 @@ pub(super) fn register_screen_and_window_props(ctx: &mut Context, bridge: &HostB
     let b = bridge.clone();
     let dpr_getter = NativeFunction::from_copy_closure_with_captures(
         |_this, _args, bridge, _ctx| {
-            #[allow(clippy::cast_precision_loss)]
-            Ok(JsValue::from(f64::from(bridge.device_pixel_ratio())))
+            // `device_pixel_ratio()` is `f64` (the lossless winit scale, C3 R3) — a
+            // fractional DPR like 1.2 reaches JS exact, not `f32`-rounded.
+            Ok(JsValue::from(bridge.device_pixel_ratio()))
         },
         b,
     );
